lgaleana/Small fixes
